Add new module "operation"
