+++ src/user/messages.rs
@@ -4,7 +4,8 @@ use serde::{Deserialize, Serialize};
 
 use super::model::User;
 use crate::database::Connection;
-use crate::message::{MessageResponder, Operation, OperationResult};
+use crate::message::MessageResponder;
+use crate::operation::{self, Operation};
 
 #[derive(Deserialize, Serialize)]
 #[serde(tag = "type", content = "payload")]
@@ -47,7 +48,7 @@ pub mod active_authors_query {
     impl Operation for Payload {
         type Output = Vec<i32>;
 
-        async fn execute(&self, connection: Connection<'_, '_>) -> OperationResult<Self::Output> {
+        async fn execute(&self, connection: Connection<'_, '_>) -> operation::Result<Self::Output> {
             Ok(match connection {
                 Connection::Pool(pool) => User::fetch_active_authors(pool).await?,
                 Connection::Transaction(transaction) => {
@@ -69,7 +70,7 @@ pub mod active_reviewers_query {
     impl Operation for Payload {
         type Output = Vec<i32>;
 
-        async fn execute(&self, connection: Connection<'_, '_>) -> OperationResult<Self::Output> {
+        async fn execute(&self, connection: Connection<'_, '_>) -> operation::Result<Self::Output> {
             Ok(match connection {
                 Connection::Pool(pool) => User::fetch_active_reviewers(pool).await?,
                 Connection::Transaction(transaction) => {
@@ -101,7 +102,7 @@ pub mod activity_by_type_query {
     impl Operation for Payload {
         type Output = Output;
 
-        async fn execute(&self, connection: Connection<'_, '_>) -> OperationResult<Self::Output> {
+        async fn execute(&self, connection: Connection<'_, '_>) -> operation::Result<Self::Output> {
             Ok(match connection {
                 Connection::Pool(pool) => User::fetch_activity_by_type(self.user_id, pool).await?,
                 Connection::Transaction(transaction) => {
